Add files via upload
My rough idea for the progress bar screen. On the left you search for the students, they appear underneath in th ebox below. You click them it puts there photo in the middle then they get to choose how confident they are with the subject. The progress bar underneath could be an average of all students to get an idea for the class.